object-based-system-raft/room-svc/requirements.txt
grpcio
grpcio-tools
protobuf
